Move all db functions to model.js
Remove db funcs that were not being used
